requirements.txt
lxml>=4.9.2
python-docx>=0.8.11
